requirements.txt
# Zeno API — Python 3.11+
python-dotenv>=1.0.0
fastapi>=0.115.0,<0.116
uvicorn[standard]>=0.32.0,<0.33

# Optional: for health checks when env is set
asyncpg>=0.30.0
redis>=5.0.0
pymongo>=4.0.0
boto3>=1.35.0
blake3>=1.0.0

# Auth
bcrypt>=4.0.0
PyJWT>=2.9.0
python-multipart>=0.0.9

# Test
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
